Pure file split of `userspace-dp/src/afxdp/umem/mod.rs` (1648 LOC → 1090 LOC). Telemetry rendering moves to two sibling files so `mod.rs` is about UMEM frame management again.
- `umem/snapshot.rs` (295 LOC): impl `BindingLiveState` block
  containing `snapshot()` (the 270-LOC `BindingLiveSnapshot` struct
  literal that loads ~120 atomics under `Relaxed`) plus its private
  helper `snapshot_hist`. Visibility tightened from the mod.rs-era
  `pub(super)` (effectively `pub(in crate::afxdp)` because mod.rs's
  `super` IS `crate::afxdp`) to the spelled-out `pub(in
  crate::afxdp)` — same effective reach, matches the existing
  `mmap::MmapArea` / `profile::Owner|Peer` precedent.

- `umem/debug_state.rs` (309 LOC): `update_binding_debug_state`,
  `update_binding_idle_debug_state`, `publish_binding_debug_state`
  (private), `advance_debug_state_publish_counter`,
  `idle_debug_state_publish_due`, `flush_v_min_scratches_into`,
  plus the two cadence constants `DEBUG_STATE_PUBLISH_MASK` and
  `IDLE_DEBUG_STATE_PUBLISH_INTERVAL_NS`. All `pub(in crate::afxdp)`
  for the same reason. `publish_binding_debug_state` stays private
  because no caller outside the file invokes it directly.

- `umem/mod.rs`: gains `mod snapshot; mod debug_state;` plus
  `pub(in crate::afxdp) use debug_state::{...}` re-exporting every
  moved item the rest of `crate::afxdp` references. The
  `use crate::afxdp::worker::WorkerTimers;` import stays at the top
  of `mod.rs` because `umem/tests.rs` references `WorkerTimers`
  through `use super::*;` at `tests.rs:1384`. The E0364 precedent
  documented at `userspace-dp/src/afxdp/tx/mod.rs:38` (`pub(super)
  use` of `pub(super)` child triggers E0364) is honoured by
  declaring all child items `pub(in crate::afxdp)` directly.

Validation:

- `TMPDIR=/dev/shm CARGO_TARGET_DIR=/dev/shm/cargo cargo build
  --release`: clean.
- `cargo test --release`: 1417 unit tests pass. The pre-existing
  `snat_contract_doc_guard` integration-test failure is also
  present on origin/master (580439fa) — unrelated to this change.
- 5/5 flake check on
  `idle_debug_state_publish_cadence_is_wall_clock_based`.
- Go suite (`go test ./...`): clean.

Plan + review history:

- Plan v1 (`61ce6204`): AGY round-1
  (`review-mpmyqhc7-hrmtnh`) PLAN-NEEDS-MINOR — flagged constants
  visibility; Codex round-1 (`task-mpmypmng-c4iun8`) infra-blocked.
- Plan v2 (`07808a98`): AGY round-2
  (`review-mpmyw9wc-ehuok7`) PLAN-READY; Codex round-2
  (`task-mpmyw1lc-b5jxjk`) PLAN-NEEDS-MAJOR — caught the
  `pub(super)` vs `pub(in crate::afxdp)` bug, the
  missing-coordinator/worker/tx-dispatch caller audit, and the
  cold-path classification overreach.
- Plan v3 (`eca9aaec`): AGY round-3
  (`review-mpmz5lcl-760i8j`) PLAN-READY; Codex round-3 on the
  full plan infra-blocked (`task-mpmz5gxe-dituof`,
  `task-mpn11slm-kqu3vx`); Codex on inline-pasted plan summary
  (`task-mpn19bsi-hva4zi`) PLAN-NEEDS-MINOR — asked for explicit
  `pub(in crate::afxdp) fn snapshot`, which v3 already contains
  at line 154. Codex-stuck 3-of-4 exception applies (sandbox
  unified-exec broken across 4 attempts in this session).

Wave-2 rules: per-PR smoke skipped; `<!-- AWAITING-BATCH-MERGE -->`
marker after 4-of-4 reviewer attestation.
